% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{Betah_SBS96_GRCh37_v3.4}
\alias{Betah_SBS96_GRCh37_v3.4}
\title{SBS96 informative prior parameters for GRCh37 v3.4}
\format{
A vector of length 96
}
\usage{
Betah_SBS96_GRCh37_v3.4
}
\description{
Beta parameters for 96 single base substitution (SBS) mutational signatures
from COSMIC v3.4 using the GRCh37 genome reference.
}
\keyword{datasets}
