% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{COSMIC_v3.4_SBS96_GRCh37}
\alias{COSMIC_v3.4_SBS96_GRCh37}
\title{COSMIC v3.4 SBS96 Mutational Signatures for GRCh37}
\format{
A data frame with 96 rows and 86 columns.
}
\source{
\url{https://cancer.sanger.ac.uk/signatures/downloads/}
}
\usage{
COSMIC_v3.4_SBS96_GRCh37
}
\description{
SBS96 signatures from the COSMIC database, v3.4, for GRCh37.
}
\keyword{datasets}
