% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_PosteriorCI}
\alias{get_PosteriorCI}
\title{Posterior credible intervals for the quantities}
\usage{
get_PosteriorCI(chain)
}
\arguments{
\item{chain}{Array containg the MCMC samples.}
}
\description{
Posterior credible intervals for the quantities
}
