% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTotalMutations}
\alias{getTotalMutations}
\title{Function to obtain the aggregate mutation matrix for NMF}
\usage{
getTotalMutations(gr_Mutations)
}
\arguments{
\item{gr_Mutations}{A genomic ranges object containing the mutations,
the samples, and the coordinates.}
}
\description{
Function to obtain the aggregate mutation matrix for NMF
}
