% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_to_RefSigs}
\alias{match_to_RefSigs}
\title{Function to match the signatures with a reference set. Defaults to cosmic}
\usage{
match_to_RefSigs(sigs, ref = COSMIC_v3.4_SBS96_GRCh37)
}
\arguments{
\item{sigs}{Matrix of estimated signatures}

\item{ref}{Matrix of reference signatures}
}
\description{
Function to match the signatures with a reference set. Defaults to cosmic
}
