% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_Betas}
\alias{plot_Betas}
\title{Plot the regression coefficient matrix from the Poisson process factorization}
\usage{
plot_Betas(Betas, lowCI = NULL, highCI = NULL)
}
\arguments{
\item{Betas}{Betas matrix}

\item{lowCI}{Credible interval for the Betas matrix}

\item{highCI}{Credible interval for the Betas matrix}
}
\description{
Plot the regression coefficient matrix from the Poisson process factorization
}
