% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_Signatures}
\alias{plot_Signatures}
\title{Plot the matrix of mutational signatures}
\usage{
plot_Signatures(
  signatures,
  lowCI = NULL,
  highCI = NULL,
  palette = c("#40BDEE", "#020202", "#E52925", "#CCC9CA", "#A3CF62", "#ECC5C5")
)
}
\arguments{
\item{signatures}{Matrix of signature to plot}

\item{lowCI}{Matrix containing the higher end of the posterior credible interval}

\item{highCI}{Matrix containing the higher end of the posterior credible interval}

\item{palette}{Palette for the substitutions. Defaults to the values in COSMIC}
}
\description{
Plot the matrix of mutational signatures
}
