% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_ThetaBaseline}
\alias{plot_ThetaBaseline}
\title{Plot the matrix of baseline values}
\usage{
plot_ThetaBaseline(Theta, clust = NULL, nclust = 10)
}
\arguments{
\item{Theta}{Matrix of baseline activities}

\item{clust}{Cluster profiles of the observations.}

\item{nclust}{Number of clusters to split the data.
Default is 10 for visualization. This is valid only if \code{clust = NULL}}
}
\description{
Plot the matrix of baseline values
}
