% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigPoisProcess.R
\name{sample_signatures}
\alias{sample_signatures}
\title{Sample a mutational signature matrix}
\usage{
sample_signatures(Alpha)
}
\arguments{
\item{Alpha}{Matrix of dimension I x K.}
}
\description{
Sample a mutational signature matrix
}
