% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigPoisProcess.R
\name{SigPoisProcess.controls}
\alias{SigPoisProcess.controls}
\title{Control values for the estimation algorithm}
\usage{
SigPoisProcess.controls(
  nsamples = 2000,
  burnin = 2000,
  maxiter = 5000,
  n_iter_betas = 2,
  tol = 1e-06
)
}
\arguments{
\item{nsamples}{Number of mcmc iterations}

\item{burnin}{Number of burnin iterations}

\item{maxiter}{Maximum number of iteratiors for the optimization}

\item{n_iter_betas}{Number of fisher scoring per iteration in the optimization}

\item{tol}{Tolerance for convergence}
}
\description{
Control values for the estimation algorithm
}
