% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigPoisProcess.R
\name{SigPoisProcess.init}
\alias{SigPoisProcess.init}
\title{Initial values for the algorithms}
\usage{
SigPoisProcess.init(
  R_start = NULL,
  Theta_start = NULL,
  Betas_start = NULL,
  Mu_start = NULL,
  Sigma2_start = NULL
)
}
\arguments{
\item{R_start}{Start value for the signature matrix.}

\item{Theta_start}{Start value for the baseline  matrix.}

\item{Betas_start}{Start value for the regression coefficient matrix.}

\item{Mu_start}{Start value for the relevance weights.}

\item{Sigma2_start}{Start value for the variance of the regression coefficients.}
}
\description{
Initial values for the algorithms
}
