% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigPoisProcess.R
\name{SigPoisProcess.PriorParams}
\alias{SigPoisProcess.PriorParams}
\title{Prior parameters in the model}
\usage{
SigPoisProcess.PriorParams(
  a = 1.01,
  alpha = 1.01,
  epsilon = 0.001,
  a0 = NULL,
  b0 = NULL,
  c0 = 100,
  d0 = 1,
  SigPrior = NULL
)
}
\arguments{
\item{a}{Prior parameter in the baseline coefficients (gamma prior).}

\item{alpha}{Prior parameter for the signature (Dirichlet prior).}

\item{epsilon}{Defaults to 0.001.}

\item{a0}{Prior for mu ~ InvGamma(a0, b0). Defaults to NULL (compressive).}

\item{b0}{Prior for mu ~ InvGamma(a0, b0). Defaults to NULL (compressive).}

\item{c0}{Prior over sigma2 ~ InvGamma(c0, d0).}

\item{d0}{sigma2 ~ InvGamma(c0, d0).}

\item{SigPrior}{Informative prior matrix for the Dirichlet.}
}
\description{
Prior parameters in the model
}
