% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigPoisProcess.R
\name{SigPoisProcess}
\alias{SigPoisProcess}
\title{Poisson process factorization for mutaitonal signatures analysis}
\usage{
SigPoisProcess(
  Mutations,
  X,
  X_total,
  method = "gibbs",
  K = 10,
  prior_params = list(),
  controls = list(),
  init = list()
)
}
\arguments{
\item{Mutations}{Dataframe containing the mutations}

\item{X}{Covariates}

\item{X_total}{Total surface of the Poisson process}

\item{method}{Which method to use for the estimation. Available options are 'gibbs', 'cavi' and 'map'}

\item{prior_params}{The hyperparameters of the prior.}

\item{controls}{Control parameters}
}
\value{
An object
}
\description{
Poisson process factorization for mutaitonal signatures analysis
}
