% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigPoisProcess_mult.R
\name{SigPoisProcess_mult}
\alias{SigPoisProcess_mult}
\title{Poisson process factorization for topographic-dependent mutational signatures analysis}
\usage{
SigPoisProcess_mult(
  gr_Mutations,
  SignalTrack,
  bin_weight,
  K = 10,
  method = "map",
  prior_params = SigPoisProcess_mult.PriorParams(),
  controls = SigPoisProcess_mult.controls(),
  init = SigPoisProcess_mult.init(),
  update_Signatures = TRUE,
  update_Theta = TRUE,
  update_Betas = TRUE
)
}
\arguments{
\item{gr_Mutations}{GenomicRanges object contatining covariates and samples}

\item{SignalTrack}{Matrix containing signals for the whole genome}

\item{bin_weight}{Weights to give to each region}

\item{K}{Upper bound to the number of signatures expected in the data}

\item{method}{Which method to use for the estimation. Available options are 'mle' and 'map'}

\item{prior_params}{The hyperparameters of the prior.}

\item{controls}{Control parameters for each method. See \code{SigPoisProcess.contol}}

\item{init}{Initialization parameters. See \code{SigPoisProcess.init}}
}
\value{
An object of class SigPoisProcess_mult
}
\description{
Poisson process factorization for topographic-dependent mutational signatures analysis
}
