% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigPoisProcess.R
\name{SigPoisProcess}
\alias{SigPoisProcess}
\title{Poisson process factorization for topographic-dependent mutational signatures analysis}
\usage{
SigPoisProcess(
  gr_Mutations,
  df_areas,
  K = 10,
  method = "map",
  compressType = "sig_cov",
  prior_params = SigPoisProcess.PriorParams(),
  controls = SigPoisProcess.controls(),
  init = SigPoisProcess.init()
)
}
\arguments{
\item{gr_Mutations}{GenomicRanges object contatining covariates and samples}

\item{df_areas}{data.frame containing the total areas of each covariate for each sample}

\item{K}{Upper bound to the number of signatures expected in the data}

\item{method}{Which method to use for the estimation. Available options are 'mle', 'map' and 'gibbs'}

\item{compressType}{Type of penalty for the automatic selection of priors using compressive hyperpriors. Available type are 'sig' and 'sig_cov'}

\item{prior_params}{The hyperparameters of the prior.}

\item{controls}{Control parameters for each method. See \code{SigPoisProcess.contol}}

\item{init}{Initialization parameters. See \code{SigPoisProcess.init}}
}
\value{
An object of class SigPoisProcess.
}
\description{
Poisson process factorization for topographic-dependent mutational signatures analysis
}
